requirements.txt
Flask==1.1.1
Flask-WTF==0.14.2
WTForms==2.2.1
Jinja2==2.11.1
requests==2.24.0
werkzeug==0.16.1
requests
azure-functions